idbo/requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.1.0
boto3==1.28.25
botocore==1.31.25
celery==5.3.4
certifi==2023.7.22
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==4.2.4
django-admin-rangefilter==0.11.1
django-dynamic-raw-id==3.0
django-minio-backend==3.5.0
django-minio-storage==0.5.6
django-storages==1.13.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.7
et-xmlfile==1.1.0
gunicorn==21.2.0
inflection==0.5.1
jmespath==1.0.1
kombu==5.3.2
minio==7.1.15
numpy==1.25.2
openpyxl==3.1.2
packaging==23.1
pandas==2.0.3
pdfrw==0.4
Pillow==10.0.0
prompt-toolkit==3.0.39
psycopg2==2.9.7
PyJWT==2.8.0
PyPDFForm==1.3.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
redis==5.0.0
reportlab==4.0.4
s3transfer==0.6.1
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
vine==5.0.0
wcwidth==0.2.6
